Unify the ledger service
