+++ node/Cargo.toml
@@ -89,7 +89,6 @@ version = "=2.1.7"
 [dependencies.snarkos-node-sync]
 path = "./sync"
 version = "=2.1.7"
-features = [ "ledger" ]
 
 [dependencies.snarkos-node-tcp]
 path = "./tcp"
