+++ node/sync/Cargo.toml
@@ -18,8 +18,7 @@ edition = "2021"
 
 [features]
 default = [ ]
-ledger = [ "snarkos-node-sync-ledger-service/ledger" ]
-test = [ "snarkos-node-sync-ledger-service/test", "snarkos-node-sync-locators/test" ]
+test = [ "snarkos-node-sync-locators/test" ]
 
 [dependencies.anyhow]
 version = "1.0"
@@ -43,12 +42,12 @@ version = "0.8"
 [dependencies.serde]
 version = "1"
 
-[dependencies.snarkos-node-sync-communication-service]
-path = "communication-service"
+[dependencies.snarkos-node-narwhal-ledger-service]
+path = "../narwhal/ledger-service"
 version = "=2.1.7"
 
-[dependencies.snarkos-node-sync-ledger-service]
-path = "ledger-service"
+[dependencies.snarkos-node-sync-communication-service]
+path = "communication-service"
 version = "=2.1.7"
 
 [dependencies.snarkos-node-sync-locators]
@@ -64,10 +63,6 @@ features = [ "rt", "signal" ]
 
 [dependencies.tracing]
 version = "0.1"
-
-[dev-dependencies.snarkos-node-sync-ledger-service]
-path = "ledger-service"
-features = [ "test" ]
 
 [dev-dependencies.snarkos-node-sync-locators]
 path = "locators"
