--- node/sync/ledger-service/Cargo.toml
@@ -1,33 +1,0 @@
-[package]
-name = "snarkos-node-sync-ledger-service"
-version = "2.1.7"
-authors = [ "The Aleo Team <hello@aleo.org>" ]
-description = "A ledger service for the sync module in a decentralized operating system"
-homepage = "https://aleo.org"
-repository = "https://github.com/AleoHQ/snarkOS"
-keywords = [
-  "aleo",
-  "cryptography",
-  "blockchain",
-  "decentralized",
-  "zero-knowledge"
-]
-categories = [ "cryptography", "operating-systems" ]
-license = "Apache-2.0"
-edition = "2021"
-
-[features]
-default = [ ]
-ledger = [ ]
-mock = [ "parking_lot" ]
-test = [ "mock" ]
-
-[dependencies.async-trait]
-version = "0.1"
-
-[dependencies.snarkvm]
-workspace = true
-
-[dependencies.parking_lot]
-version = "0.12"
-optional = true
