--- node/sync/ledger-service/README.md
@@ -1,7 +1,0 @@
-# snarkos-node-sync-ledger-service
-
-[![Crates.io](https://img.shields.io/crates/v/snarkos-node-sync-ledger-service.svg?color=neon)](https://crates.io/crates/snarkos-node-sync-ledger-service)
-[![Authors](https://img.shields.io/badge/authors-Aleo-orange.svg)](https://aleo.org)
-[![License](https://img.shields.io/badge/License-Apache%202.0-blue.svg)](./LICENSE.md)
-
-The `snarkos-node-sync-ledger-service` crate provides a ledger service implementation for a synchronization module.
